requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0

# LangChain & AI
langchain==0.1.0
langchain-openai==0.0.2
openai>=1.0.0
tiktoken>=0.5.0

# Vector Database
pinecone-client>=3.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8

# Data Processing
pandas==2.1.3
numpy==1.26.2

# HTTP Client
requests==2.31.0
httpx==0.25.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication
sqlalchemy==2.0.23
bcrypt==4.1.2
email-validator==2.1.0

# Logging & Monitoring
loguru==0.7.2

